requirements.txt
fastapi==0.115.6
uvicorn==0.29.0
spacy==3.8.2
rich==13.7.1
pdfplumber==0.11.4
jinja2==3.1.4
Flask==2.3.3
openai==1.3.5
python-dotenv==1.0.0